% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wg_util.R
\name{calc_vertex}
\alias{calc_vertex}
\title{Calculate the vertex of a parabola defined by a quadratic function}
\usage{
calc_vertex(a, b, c)
}
\arguments{
\item{a}{numeric coefficient for the squared term}

\item{b}{numeric coefficient for the 1st order term}

\item{c}{numeric intercept}
}
\value{
named vector with the x and y coordinates of the vertex
}
\description{
Calculates the vertex for f(x) = ax^2 + bx + c according to the
 formula:

 vertex x = -b/2a

 vertex y = f(vertex x)
}
