% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref_to_hec.R
\name{gridref_to_hec}
\alias{gridref_to_hec}
\title{Convert a grid ref (letter form) of any precision to hectad (3 characters)}
\usage{
gridref_to_hec(gr = NULL)
}
\arguments{
\item{gr}{character vector giving the grid references (e.g. "B1641", "B161415)}
}
\description{
Convert a grid ref (letter form) of any precision to hectad (3 characters)
}
