% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lat_long_to_OSI.R
\name{lat_lon_to_osi}
\alias{lat_lon_to_osi}
\title{Convert WGS84 lat/long to OSI TM75 eastings/northings.}
\usage{
lat_lon_to_osi(x, lat = "Latitude", lon = "Longitude", orig_crs = NULL)
}
\arguments{
\item{x}{A data frame containing latitude and longitude coordinates.}

\item{lat, lon}{Character string giving the names of the latitude and
longitude columns.}

\item{orig_crs}{A character string giving the proj4string to pass to
\code{\link[sp]{CRS}}.}
}
\description{
Convert latitude/longitude to eastings and northings on the Irish OSI TM75
grid.
}
\details{
This function calls \code{\link[sp]{spTransform}} to convert
latitude/longitude coordinates from WGS84 to the Irish OSI TM75 grid.  The
main purpose of this function is to make this slightly more automatic for
Willson's PhD work by providing the destination epsg:29903 code in the
function.  The function also re-names the output coordinate fields to
'eastings' and 'northings'.
}
\examples{
set.seed(500)
df <- data.frame(site = c("A", "B", "C", "D", "E"),
                 Latitude = rnorm(mean = 54.3, sd = 1, 5),
                 Longitude = rnorm(mean = -6, sd = 0.5, 5))
df <- lat_lon_to_osi(df, orig_crs = "+init=epsg:4326")
plot(df$Latitude ~ df$Longitude)
plot(df$northings ~ df$eastings)

}
\seealso{
\code{\link[sp]{spTransform}}, \code{\link[sp]{CRS}},
\code{\link{rgdal}}, \code{\link[sp]{proj4string}},
\url{http://spatialreference.org/ref/epsg/tm75-irish-grid/}
}
