% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wg_util.R
\name{logistic}
\alias{logistic}
\title{Calculate logistic of a number}
\usage{
logistic(x)
}
\arguments{
\item{x}{a numeric value}
}
\value{
the logistic of x
}
\description{
Calculates logistic(x) according to the formula:
1 / (1 + exp(-x))
}
