% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wg_util.R
\name{make_name}
\alias{make_name}
\title{Create names by assigning a prefix and a number}
\usage{
make_name(suffix, prefix)
}
\arguments{
\item{suffix}{character string or integer to be used in suffix of name}

\item{prefix}{character string type identifier (e.g. "hec" or "site")
(integer won't work to start character string name)}
}
\description{
Create names by assigning a prefix and a number
}
