% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wg_util.R
\name{pa}
\alias{pa}
\title{Convert to presence / absence}
\usage{
pa(x)
}
\arguments{
\item{x}{a dataframe or vector with numeric values to be converted to 0 or 1.}
}
\description{
Convert to presence / absence
}
