% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResponse_ggplot2.R
\name{plotResponse_ggplot2}
\alias{plotResponse_ggplot2}
\title{Plot species probability of occurrence along with the response curve for
each of 2 predictor variables in marginal plots.  This}
\usage{
plotResponse_ggplot2(df, pred_1, pred_2, resp, coefs)
}
\arguments{
\item{df}{data frame holding values for all predictor variables, principal
components, and responses (logit(p) and probability)
for each site in the raster}

\item{pred_1}{character string giving column name of predictor variable 1}

\item{pred_2}{character string giving column name of predictor variable 2}

\item{resp}{- character string giving column name of the response variable
(e.g. probability of occurrence)}

\item{coefs}{- a named list giving a numeric vector with the coefficients
 for the linear predictors, including the intercept and first and second
 degree polynomials for each of the
 predictor variables. The name is the same as the name of the predictor
 variable, and the values are numeric coefficients.
@param axis1_offset - the value that is subtracted from the values in the
 raster of the first predictor before squaring that predictor.
@param axis2_offset - the value subtracted from the values in the raster of
 the second predictor before squaring that predictor.}
}
\value{
This function prints to the graphics device rather than returning an
  object.
}
\description{
For the \code{coefs} argument, use an argument like
\code{coefs = list(elevation = c(0.5, 0.1, 0.3))} for a response of the form
logit(prob_of_occurrence) = 0.5 + 0.1*elevation + 0.3*(elevation^2).
Use a \code{0} for the
coefficient of the polynomial term to produce a straght-line response.
}
\examples{

}
\seealso{
my code is based on this:
\url{https://stackoverflow.com/questions/17370853/align-ggplot2-plots-vertically/17371177#17371177}
}
