% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_envelope.R
\name{quantile_envelope}
\alias{quantile_envelope}
\title{Calculate quantile envelopes}
\usage{
quantile_envelope(x, upper = 0.975, lower = 0.025, n.eval.points,
  type = "density")
}
\arguments{
\item{x}{a dataframe with the observed values (column must be named
"observed") and a column for each replicate to be used for quantile
calculation}

\item{upper}{a numeric value giving the desired upper quantile boundary (0
to 1)}

\item{lower}{a numeric value giving the desired lower quantile boundary (0
to 1)}

\item{n.eval.points}{integer giving the number of points at which to compute
pointwise quantiles.  This is passed to \code{\link[ggplot2]{geom_density}}.}

\item{type}{character vector indicating whether to produce envelope values
as a density ("density") or expected value ("count")}
}
\value{
a data frame with three columns holding the
  observed values, the upper envelope values, and the lower envelope values
}
\description{
Calculate the pointwise quantile envelopes for comparison to observed
data.
}
\details{
TODO: figure out difference b/t pointwise and complete envelopes

TODO: define a plot.env.envelope method
}
\examples{
# make some example data
obs <- c(1:5, 7:8, 11:12, 19:25)
dat <- data.frame(observed = obs, s1 = sample(obs, replace = T),
                   s2 = sample(obs, replace = T),
                   s3 = sample(obs, replace = T),
                   s4 = sample(obs, replace = T),
                   s5 = sample(obs, replace = T))
# calculate the pointwise 95\% quantile envelopes
result <- quantile_envelope(x = dat, n.eval.points = 20,
                            upper = 0.975, lower = 0.025)
plot(result$observed ~ result$x, type = "l", col = "black",
ylim = c(min(result$low_bnd), max(result$hi_bnd)),
main = "Quantile Envelope", ylab = "Density")
lines(result$low_bnd ~ result$x, col = "blue")
lines(result$hi_bnd ~ result$x, col = "blue")

}
\seealso{
\code{\link[ggplot2]{geom_density}}
}
