\name{wgutil-package}
\alias{wgutil-package}
\alias{wgutil}
\docType{package}
\title{
\packageTitle{wgutil}
}
\description{
 \packageDescription{wgutil}
}
\details{
 This package is an assortment of functions that I frequently find useful. 
 
 The DESCRIPTION file:
 \packageDESCRIPTION{wgutil}
 \packageIndices{wgutil}

}
\author{
 \packageAuthor{wgutil} 
 
 Maintainer: \packageMaintainer{wgutil}
}
\references{
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
