\name{wgutil-package}
\alias{wgutil-package}
\alias{wgutil}
\docType{package}
\title{
\packageTitle{wgutil}
}
\description{
\packageDescription{wgutil}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{wgutil}
\packageIndices{wgutil}
~~ An overview of how to use the package, including the most important functions ~~
 This package is an assortment of functions that I frequently find useful. 
}
\author{
\packageAuthor{wgutil}

Maintainer: \packageMaintainer{wgutil}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
